incident-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
pytest
httpx
python-dotenv